mapping optimization

+++ src/main/java/com/rathon/manatee/database/model/Company.java
@@ -9,6 +9,6 @@ public class Company {
     private Long id;
     private String name;
     private String address;
-    private int industryId;
+    private Long industryId;
     private String registrationNumber;
 }
